# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "E-commerce API",
        "description": "API REST segura para gestión de productos, órdenes y pagos mediante Stripe.",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/orders/{order}/payment": {
            "post": {
                "tags": [
                    "Payments"
                ],
                "summary": "Create a Stripe PaymentIntent for an order",
                "operationId": "6edc2f064355a1270b551657b5af0f47",
                "parameters": [
                    {
                        "name": "order",
                        "in": "path",
                        "description": "Order ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 3
                    }
                ],
                "responses": {
                    "200": {
                        "description": "PaymentIntent created successfully"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "User is not allowed to pay this order"
                    },
                    "422": {
                        "description": "Order cannot be paid"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/stripe/webhook": {
            "post": {
                "tags": [
                    "Payments"
                ],
                "summary": "Receive Stripe webhook events",
                "operationId": "7ec111ffa3e46ef4373aadcb780daef0",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "description": "Stripe webhook event payload",
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Webhook received successfully"
                    },
                    "400": {
                        "description": "Invalid payload or signature"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}